Fix disabled login/register buttons after failed attempt
In the login window, when the user enter a wrong email or password, the
"Log into my ... account" button used to remain disabled, which was very
confusing for users.

Steps to reproduce:
- Go to /login
- Type wrong email/password combination
- Press "Log into..." button

Similarly, users could not click the "register button after a failed
register attempt. For example, this occurred to users with an invalid
email address, such as "fooatgmaildotcom".